Revert "TST Uses float32"
This reverts commit dfba2c48c04217b7000a16b74ee3cd0d58859642.

+++ sklearn/cluster/optics_.py
@@ -233,7 +233,7 @@ if metric=’precomputed’.
         self : instance of OPTICS
             The instance.
         """
-        X = check_array(X, dtype=np.float32)
+        X = check_array(X, dtype=np.float)
 
         if self.cluster_method not in ['dbscan', 'xi']:
             raise ValueError("cluster_method should be one of"
